move url parsing to a class
